Reconcile opening campaign, economy and Rustboro Gym with the v4 book
Synchronize approved team reviews through battle 28, restore Gina/Mia, enforce six-member Gym parties, and give Roxanne a coherent Mega Aerodactyl Tailwind plan with a native reveal and story explanation. Preserve provisional difficulty levels and the paused C15 Dwayne experiment.

Include outstanding shared AI, trainer-level, stage-legal preparation and early economy fixes with focused native regressions. Refresh README, current handoff, checkpoint and verification guidance; distinguish current source from older published artifacts.

+++ include/difficulty.h
@@ -9,7 +9,7 @@ struct Pokemon;
 enum DifficultyLevel GetCurrentDifficultyLevel(void);
 void SetCurrentDifficultyLevel(enum DifficultyLevel);
 u8 GetTrainerLevelReduction(void);
-u8 GetCampaignTrainerLevel(s8 offset);
+u8 GetCampaignTrainerLevel(s16 offset);
 
 enum DifficultyLevel GetBattlePartnerDifficultyLevel(u16);
 enum DifficultyLevel GetTrainerDifficultyLevel(u16);
